TestFileUtils: reunite tests now that they are all no-mock tests
